Add files via upload
Added basic obstacles to world generation. Added script to randomly place obstacles and humans correctly. The main supervisor will need to be adjusted to account for the pre-placed humans (I haven't made any changes as I don't fully understand what it does and to prevent conflicts) and in the newGeneratedWorld the old supervisor is disabled (and will crash) with the new one only to place the items.